requirements/base.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-appconf==1.0.3
django-compressor==2.3
django-libsass==0.7
django-tinymce==2.8.0
gunicorn==19.9.0
idna==2.8
libsass==0.19.3
Pillow==6.2.1
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
stripe==2.40.0
urllib3==1.25.7
whitenoise==4.1.4
django-storages
boto